% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{bathymetry}
\alias{bathymetry}
\title{Bathymetric color palettes with tiepoints}
\format{
A \code{list} with 1 \code{data.frame} elements:
\describe{
\item{\code{sandy}}{: A color ramp resembling a sandy beach. }
}
}
\usage{
data(bathymetry)
}
\description{
The object contains \code{data.frame}-class objects to be used with the \code{\link{expand}} function to produce full calibrated color ramps.
}
\examples{
data(bathymetry)
sandyExp <- expand(bathymetry$sandy, n=200)
plot(sandyExp)
}
\keyword{datasets}
