% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{colorpoints}
\alias{colorpoints}
\title{Colored points}
\usage{
colorpoints(
  x,
  y = NULL,
  z,
  ramp = NULL,
  col = NULL,
  breaks = NULL,
  legend = list(x = "topleft"),
  ...
)
}
\arguments{
\item{x}{The x argument of \code{points}.}

\item{y}{The y argument of \code{points}.}

\item{ramp}{A calibramp object (including both breaks and colors).}

\item{col}{A vector of colors. Used only if ramp is not given.}

\item{breaks}{A vector of breaks. If given, this has to be one element longer than the length of col.}

\item{legend}{A list of arguments passed to the ramplegend function. Set to \code{NULL} if you do not want to plot a legend.}

\item{...}{Arguments passed to \code{points}.}
}
\description{
The function is a wrapper around the points function, controling the color of the points similar to \code{ggplot} but using S-style plotting.
}
