% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{expand}
\alias{expand}
\title{Create color ramp palette from fixed, color-tiepoint objects}
\usage{
expand(x, n)
}
\arguments{
\item{x}{A \code{data.frame} object with two columns: \code{color} for hexadecimal color values, \code{z} for their position.}

\item{n}{Single integer number}
}
\value{
A \code{list}-class object, with three elements: \code{$col} hexadecimal color values, \code{$mid} z-values of midpoints (one for every color, \code{$breaks} separator borders between color values.
}
\description{
Function to create ramp palettes from fixed color positions
}
\examples{
data(topos)
ramp <- expand(topos$havanna2, n=200)
plot(NULL, NULL, xlim=c(-1,1), ylim=c(-8000,5000))
rect(xleft=-2, xright=2, ybottom=ramp$breaks[-1], ytop=ramp$breaks[-length(ramp$breaks)],
col=ramp$col, border=NA)
}
