% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{paleomap}
\alias{paleomap}
\title{Topographic gradient map color map of the PALEOMAP project}
\format{
A \code{calibramp}-class \code{list} with 3 \code{numeric}s:
\describe{
\item{\code{col}}{: The color levels as hexadecimal RGB values. }
\item{\code{breaks}}{: The boundaries for the individual levels.}
\item{\code{mid}}{: The mid values of the color levels.}
}
}
\source{
\url{https://zenodo.org/records/10659112}
}
\usage{
data(paleomap)
}
\description{
The elevation to color bindings are the work of C. Scotese. Data from Scotese, C. R., Vérard, C., Burgener, L., Elling, R. P., & Kocsis, A. T. (2025). The Cretaceous World: Plate Tectonics, Paleogeography, and Paleoclimate. Geological Society, London, Special Publications, 544(1), SP544–2024..
}
\keyword{datasets}
