% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rampage}
\alias{rampage-package}
\alias{rampage}
\title{Strechable Color Ramps}
\description{
Value-calibrated color ramps can be useful to emphasize patterns in weirdly distributed data. Tiepoints can be used to fix colors to specific values, and the resulting association can then be expanded into full color ramps. Such ramps can be used a variety of cases when heatmap type plotting is necessary, including the visualization of raster and polygon spatial data.
}
\details{
This is still a Beta version. As is R, this is free software and comes with ABSOLUTELY NO WARRANTY. Nevertheless, notes about found bugs and suggestions are more than welcome.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adamtkocsis.com/rampage/}
  \item Report bugs at \url{https://github.com/adamkocsis/rampage/issues}
}

}
\author{
Adam T. Kocsis (adam.t.kocsis@gmail.com)
}
