% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ramplegend}
\alias{ramplegend}
\title{Create a heatmap legend based on calibrated color ramp values}
\usage{
ramplegend(
  x = "topleft",
  y = NULL,
  shift = c(0, 0),
  ramp = NULL,
  col = NULL,
  breaks = NULL,
  zlim = NULL,
  height = 3,
  width = 0.3,
  tick.length = 0.15,
  cex = 1,
  box.args = list(col = "#ffffffbb"),
  horizontal = FALSE,
  at = NULL,
  label = NULL
)
}
\arguments{
\item{x}{Position of the legend or the left coordinate of legend box.}

\item{y}{Coordinate of the upperleft coordinate of the legend (if needed).}

\item{shift}{Used instead of the inset argument of the default legend. If plotted within the inner plotting area, the x and y user coordinates with which the position of the legend will be shifted to be shifted.}

\item{ramp}{A calibrated color ramp object. Either \code{ramp} or both \code{col} and \code{breaks} are required.}

\item{col}{Vector of colors.}

\item{breaks}{Breaks between the colors..}

\item{height}{Height of the legend bar in inches.}

\item{width}{Width of the legend bar in inches.}

\item{tick.length}{The length of the legend's ticks.}

\item{cex}{Legend size scaler.}

\item{box.args}{the box's arguments.}

\item{horizontal}{Legend orientation. Not yet implemented!}

\item{at}{Where should the legend be drawn in the z dimension?}

\item{label}{What are the labels}
}
\value{
The function has no return value.
}
\description{
Create a heatmap legend based on calibrated color ramp values
}
\examples{
# Examples will come here
plot(1:10)
}
