% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit.R
\name{trimramp}
\alias{trimramp}
\title{Trimming a calibrated color ramp object.}
\usage{
trimramp(x, low = NULL, high = NULL)
}
\arguments{
\item{x}{A calibrated color ramp (e.g. \code{calibramp} class object.}

\item{low}{The minimum value in the calibrated ramp.}

\item{high}{The maximum value in the calibrated ramp.}
}
\value{
A trimmed version of \code{x}
}
\description{
Trimming a calibrated color ramp object.
}
\examples{
data(paleomap)
trimmed <- trimramp(paleomap, low=-5000, high=5000)
}
