% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit.R
\name{limit}
\alias{limit}
\title{Limit range of values in a SpatRaster object}
\usage{
limit(x, y = NULL, min = NULL, max = NULL)
}
\arguments{
\item{x}{object to limit, a \code{SpatRaster} object.}

\item{y}{Range, numeric vector with two values, or data frame of positioned color values.}

\item{min}{The minimum value to have in the raster.}

\item{max}{The maximum value to have in the raster.}
}
\value{
A \code{SpatRaster} object.
}
\description{
Limit range of values in a SpatRaster object
}
\examples{
# an example for spatraster limitation
}
