% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ramps}
\alias{ramps}
\alias{gradinv}
\title{Color gradient ramps}
\usage{
gradinv(n)
}
\arguments{
\item{n}{(\code{numeric}) Number of different colors to generate from the palette}
}
\value{
A function producing a color gradient ramp.
}
\description{
The object contains functions produced by the \code{\link[grDevices:colorRamp]{colorRampPalette}} function.
}
\details{
You can also view single palettes individually. The following color palettes are implemented:
\itemize{
\item \code{gradinv()}: inverse heatmap.
}
}
\examples{
cols <- gradinv(20)
plot(1:20, col=cols, pch=16, cex=2)
}
