% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{topos}
\alias{topos}
\title{Calibrated topographic color palettes}
\format{
A \code{list} with 2 \code{data.frame}s:
\describe{
\item{\code{jakarta}}{: The "Jakarta" theme.}
\item{\code{havanna2}}{: The "Havanna-2" theme.}
\item{\code{tokio1}}{: The "tokio-1" theme.}
\item{\code{zagreb}}{: The "zagreb" theme.}
}
}
\usage{
data(topos)
}
\description{
The object contains \code{data.frame}-class objects.
}
\keyword{datasets}
