% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{topos}
\alias{topos}
\title{Topographic color palettes with tiepoints}
\format{
A \code{list} with 4 \code{data.frame}s:
\describe{
\item{\code{jakarta}}{: The "Jakarta" theme, color values by \emph{Arcanographia}. }
\item{\code{havanna2}}{: The "Havanna-2" theme, color values by \emph{Arcanographia}.}
\item{\code{tokio1}}{: The "Tokio-1" theme, color values by \emph{Arcanographia}.}
\item{\code{zagreb}}{: The "Zagreb" theme, color values by \emph{Arcanographia}.}
}
}
\usage{
data(topos)
}
\description{
The object contains \code{data.frame}-class objects.
}
\keyword{datasets}
